requirements.txt
rpy2>=3.5.15
pandas>=2.1.4
pillow>=10.2.0
numpy>=1.26.3
